Use x86_64-pc-none-elf instead of x86_64-pc-linux-elf
Using the Linux target when we clearly aren't is dangerous, even if
it works for now. We don't know what other flags are inserted by
default with the Linux target, so we use a generic one instead.

To facilitate this, the linking action has been patched to use the linker
instead of the system compiler, which is the default behavior when
the target is none, not Linux.