x86_64: Fix buffer overrun in VGA console
When using two or more color specifiers (\e[3xm), the length variable
gets decremented by x - 1 times to many, where x is the number of
color specifiers because we decrease length in the loop itself.

Instead, we should only check the value in the loop conditional
and manually decrement for everything except for color specifiers.