tsc: Use different rdtsc inline assembly based on architecture
rdtsc places the value into EAX and EDX, so to read the tsc we need to
combine them into one.

On 32-bit platforms, =A will automatically merge those two together into a
variable. This is the previous behavior.

The problem occurs on 64-bit platforms, where =A refers solely to RAX.
We would end up getting a 32-bit value which wrapped around ~1.7 seconds
after boot, and manifested as skipped time values.

This commit defines tsc_read() based on architecture, with 32-bit platforms
keeping the original function and 64-bit platforms getting a function
that merges EAX and EDX manually into one variable, which resolves
the time skip issues.