Relicense to BSD 3-Clause license
I want this to be more open, and GPL has too many restrictions.
Before, I was unable to switch the license due to the presence of Linux
code in the kernel. Linux code is GPLv2, and BSD/MIT are fundamentally
incompatible with GPLv2.

However, with the completion of issue #2 which removed all the Linux
code, I am free to license as I wish as the code is all mine. To
encourage contributions to this project, I choose the BSD because
it's open yet offers strong legal protections.