Add new list implementation and convert irqchip to use as a POC
The new list implementation is based of Linux's and allows us to have
multiple list elements in one struct, which will come in handy
especially for the task management code.

To make sure the code works, the irqchip system has been converted to use
the new list implementation.

The old implementation remains in list.h, but will be removed once the
entire kernel has been converted.