Implement kthread_exit so kernel threads can actually exit properly
Before, kernel threads HAD to loop forever, because exiting the thread
(aka returning) would result in it returning to a random address
and causing problems in general.

This commit introduces kthread_exit which deschedules the thread from
the scheduler and loops until we get switched out. In the future,
ideally we would call sched_yield or something like that.

To make sure the function actually returns to kthread_exit, we
set up the stack manually so that the last value on the stack is the
address of kthread_exit, which ret then jumps to.