Store stack metadata in free memory itself instead of allocating maximum
The previous behavior had the init system allocate enough memory to store
the entire stack in memory, but that wouldn't work on large amounts
of physical memory (e.g. TB range)

This commit updates the stack to dynamically map in the physical page
into a known virtual address to access the contents of that, storing
the actual metadata in the physical page itself.

In our case, this metadata is a pointer to the next physical free page,
and we can then pop and push by simply mapping different physical
addresses into that page.

Some work needs to be done to remove x86-isms from the stack allocator,
most notably from the stack_map function, but in it's current state,
this is stable enough to replace the original stack allocation scheme.