Use symlink instead of manually copying architecture header
In the past, modifying an architecture header would require two runs:
once to copy the files, and the second time for CMake to actually
act on the changes. This is because CMake builds the dependency graph
before copying the architecture headers, so it won't know about them.

This commit changes include/arch to be a symlink to the actual arch
include folder, allowing CMake to actually be aware of the file
changes BEFORE it builds the dependency graph (because technically
it's working with the direct copy). Now, changes are picked up
the first run.