x86_64: invlpg virtual address too
We only invlpg'ed the recursive table mappings which worked OK most
of the time, but when remapping we would sometimes get invalid
TLB entries for the target virtual address.

This manifested visibly as page faults in the AHCI driver, even
though it had mapped the ABAR into memory. Checking the TLB entries
revealed that the flags were all wrong, which was a sign of
stale TLB entries.

This commit invlpg's the virtual address both times, once when
setting the recursive mapping and once when restoring the original
mapping, to ensure that the TLB entries are correct.