hashmap: Set key properly
Previously we stored the address of the key AS the key, which worked fine for
keys up to a maximum size of sizeof(void*), which was 8 bytes for x86_64 since
the key would be stored as the 'address'.

This bug manifested when we started using arrays as the keys. When storing
the key, we would store the address of the key, not the key itself, but
when retrieving, we would memcmp the key against the address stored, meaning
that it would never return true.

This commit converts bucket->key to a pointer to an array storing the actual
key. When inserting, we allocate space for the key and memcpy the key into
the array. When retrieving, we memcmp against the array and the key.